resource_rewriting: support rewritable directories
`?rr` can now be put on a link to a directory,
in which case the directory is given
a unique name instead of the files in it.
This is useful for JavaScript that e.g. dynamically selects
    ("imgs/?rr" + "piece-" + piece_number + ".png")

The directory's unique name is based on the contents of
everything in it and the things they depend on.  The
unique path of each file in it is only unique by dint of
the directory's renaming.  Unfortunately, JavaScript
[doesn't come](https://stackoverflow.com/questions/2676178/joining-relative-urls)
[with correct](http://code.google.com/p/js-uri/)
[URL joining](https://stackoverflow.com/questions/10565624/completing-relative-url-to-current-webpage),
so `("imgs/?rr" + ...)` won't trivially work without
resource-rewriting the way `"some-image.png?rr"` does.

Another small downside: every file under such directories
will be served as a resource, even if it's not used as a
resource, because it's not trivially clear *which* files
under the directory are accessed by JavaScript in this way.
So try to put the files you're accessing this way in their
own subdirectory.